Uncomment some lines in virtualship.py and move old code into package as its not old.
